shrink bookmaps even more

+++ engine/src/osfbook.rs
@@ -91,8 +91,8 @@ pub struct Book {
     pub wld_count: [i32; 61],
     pub exhausted_count: [i32; 61],
     pub common_count: [i32; 61],
-    pub symmetry_map: [&'static [i16]; 8],
-    pub inv_symmetry_map: [&'static [i16]; 8],
+    pub symmetry_map: [&'static [i8]; 8],
+    pub inv_symmetry_map: [&'static [i8]; 8],
     pub line_hash: [[[i32; 6561]; 8]; 2],
     pub book_hash_table: Vec<i32>,
     pub draw_mode: DrawMode,
@@ -103,14 +103,14 @@ pub struct Book {
     pub col_pattern: [i32; 8],
 }
 pub struct BookMaps {
-    pub b1_b1_map: [i16; 100],
-    pub g1_b1_map: [i16; 100],
-    pub g8_b1_map: [i16; 100],
-    pub b8_b1_map: [i16; 100],
-    pub a2_b1_map: [i16; 100],
-    pub a7_b1_map: [i16; 100],
-    pub h7_b1_map: [i16; 100],
-    pub h2_b1_map: [i16; 100],
+    pub b1_b1_map: [i8; 100],
+    pub g1_b1_map: [i8; 100],
+    pub g8_b1_map: [i8; 100],
+    pub b8_b1_map: [i8; 100],
+    pub a2_b1_map: [i8; 100],
+    pub a7_b1_map: [i8; 100],
+    pub h7_b1_map: [i8; 100],
+    pub h2_b1_map: [i8; 100],
 }
 pub static BOOK_MAPS: BookMaps = create_book_maps();
